Move generated CTIMER types to dedicated module
